+++ test/callbacks.jl
@@ -77,7 +77,7 @@ end
         callback.maximum_surface_wind_speed[i] = max_2norm(u_grid, v_grid)
     end
 
-    SpeedyWeather.finish!(::StormChaser, args...) = nothing
+    SpeedyWeather.finalize!(::StormChaser, args...) = nothing
 
     spectral_grid = SpectralGrid()
     callbacks = CallbackDict(NoCallback())